Extract drift inertia into useDrift hook
